tasks_service/requirements.txt
Django==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
celery==5.3.1
redis==4.6.0
requests==2.31.0
watchdog==3.0.0
newspaper3k==0.2.8
Scrapy==2.11.0